# fix(perf): Don't clone globals