@epage 2018-09-18
@emoon while there is more that can be done for performance, I think I'll be taking a break on address that after this.

Note that this means you could avoid creating a top-level `liquid::value::Object` completely by creating a struct that implements `Globals` that maintains references to the data you care about.